feat: componentização dos cards de capitulo e nivel
